[Build] Rework Makefile rules to ease addition of LOC package
This commit refactors the Makefile rules to ease the addition
of rules to integrate Line-of-Code (LOC) package in a follow-on
commit. Several redundant TEST_*BIN* etc. symbols are purged.

Existing capabilities are unchanged, and we now additionally
support:
$ make clean && CC=g++ CXX=g++ LD=g++ make all
$ make run-tests

... to build all .c, .cpp and .cc sample-programs and run them
in one-pass.

clean: rm /tmp/l3*.dat files, also.
Update `make help` usage messages to be more self-explanatory.