[Py] Optimize unpacking LOC-ID when logging Mils of logs
When L3 logging is invoked with `L3_LOC_ENABLED=1`, the unpacking
logic invokes the LOC-decoder binary for each LOC-ID found in the logs.
For sample programs that log 300 Mil logs, from the same source-code
line, this causes a noticeable slow-down in the performance of the
Py-dump script.

This commit implements a local optimization in the Py-dump script to
cache the loc-ID from each log-entry, and to skip the LOC-unpacking logic
if the `next loc-ID == prev-loc-ID`.

For the 300 Mil logging done by use-cases/single-file-C-program, on
a Linux-VM, this results in a drop in run-time from ~21s
(before this fix) to ~0.2s (with this fix). The run-time is
closer to the run-time when dumping the log-entries without
L3_LOC_ENABLED, ~0.1624s;  where the loc-ID is printed as-found,
without any unpacking.