[Perf] Develop client-server comms. program to measure L3-logging overheads
This commit simplifies the simple client-server communication program
to be usable as a performance test-bed to quantify the overheads of
L3-logging. Basic single/multiple-clients exchanging n-messages with server
that runs in serial manner (of processing messages) is working with this
commit.

The changes are as follows:

svmsg_file.h:
  - Define enum req_resp_type_t{} with few message types.
  - Specifically of interest are: REQ_MT_INIT, REQ_MT_INCR, REQ_MT_EXIT

svmsg_file_server.c
  - Simplified server. Receives messages inline from client in a forever
    loop. Currently, only implements "increment" counter.
  - Manages state of multiple client connections using simple array
  - L3-logging code is conditionally defined under L3_ENABLED. So,
    will need to build separately to invoke-and-test logging overheads.

svmsg_file_client.c
  - Sits in forever loop, sending "increment" message to server
  - Do this for n-iterations.

Makefile: Make changes to build the client/server programs with
 L3-logging OFF (baseline) and logging enabled.

test.sh: Add test-build-and-run-client-server-perf-test and include
  that in CI's build.yml . We will build-and-run a small scale run of
  this new test, with 5 concurrent clients, L3 ON and OFF.

build.yml: Add new test to run in CI in debug & release build modes.