[Perf] Makefile support to build client-server perf programs w/LOC encoding
This commit implements Makefile rules to build the client-server programs
such that L3-logging is performed with default LOC-encoding.

./test.sh : Add test-build-and-run-client-server-perf-test-l3_loc_eq_1
  which is also invoked thru a call in CI build.yml

This test-case will run the client/server app and then will invoke
the l3_dump.py to unpack the L3-log entries, with LOC-ID decoded.

As the LOC-generator, by default, spits out the LOC-decoder with
the name of the dir/ it is searching, we need a way to specify
the LOC-decoder binary to be used when dumping L3 log-entries.
l3_dump.py is now extended to support --loc-binary argument to
specify the name of the LOC-decoder binary to use. (This feature is
useful even by itself outside this very test-specific usage.)

To test this manually do:

$ make clean && CC=gcc LD=g++ BUILD_VERBOSE=1 L3_LOC_ENABLED=1 make client-server-perf-test

$ build/release/bin/use-cases/svmsg_file_server

$ build/release/bin/use-cases/svmsg_file_client

$ L3_LOC_ENABLED=1 ./l3_dump.py --log-file /tmp/l3.c-server-test.dat --binary build/release/bin/use-cases/svmsg_file_server --loc-binary build/release/bin/use-cases/client-server-msgs-perf_loc