[L3] Rework logging APIs to invoke printf() for fmt-spec validation
This commit lays down the groundwork to re-engage u-benchmarking
programs to compare the overheads of L3-logging v/s, say, logging
via printf() or write() system call. In order to be able to do that,
we first need to validate that the print format-specifiers in the
'msg' supplied to l3_log*() are consistent with the arguments supplied.
Currently, several of the use-case and sample programs are coded
loosely w/o adhering to this requirement.

This commit implements the following changes:

 - Slighly rearranges calling sequence of L3-logging APIs.
 - Defines DEBUG-version of the caller-macros, which under an
   `else if (0)` [ dead-code block ] will evaluate to a call to
   printf().
 - This will trip-up compilation errors in debug-mode builds if
   any use-case of L3-logging apis have a message with print
   format-specifiers that are inconsistent with the args supplied.

Requiring that the args-supplied are consistent with the print-
format specifiers in the `msg` also makes the job of L3-dump
utility lot more reliable and simpler. (It will also help in future
when we do support var-args to the L3-logging apis.)

Bulk of the changes are in include/l3.h with minor updates to test
programs to conform to the new calling requirements. We still
retain required 2-args to the L3-logging call. (Var-args support
will be considered in a future change.)