[Test] Add run-all-client-server-perf-tests to test.sh
This commit adds a new catch-all driver test-method, named
run-all-client-server-perf-tests, to test.sh, which can be
used to run through all client-server performance tests
that are of interest.

This new method is not going to be run in CI as all it does
is to execute other test-methods which _are_ executed in CI.

By default, all RPC-perf-tests are run with 5-concurrent clients,
each sending 1000 messages to the server.

To make it easy to run dev-test performance benchmarks on
dev-boxes, this test-method supports optional parameters:

    - number-of-messages to be sent to the server by each client
    - number-of-concurrent clients banging on the server

Usage:
  ./test.sh run-all-client-server-perf-tests [ <num-msgs> [ <num-clients> ] ]

  # Run all perf tests w/ default 1000 msgs and 5 concurrent clients.
  $ ./test.sh run-all-client-server-perf-tests

  # Run all perf tests w/ 1Mil msgs (and 5 concurrent clients)
  $ ./test.sh run-all-client-server-perf-tests $((1000 * 1000))

  # Run all perf tests w/ 1Mil msgs and a single-client
  $ ./test.sh run-all-client-server-perf-tests $((1000 * 1000)) 1