[Perf] Use -Ofast for release and -Og for debug builds.
This commits comments out the `-Ofast` flag added to Makefile,
which appears to show performance gains when client-server perf
tests are run on Linux-VM.

build.yml: Add new CI test-method, test-run-all-client-server-perf-tests,
to run this exercise for 1 Million msgs, 1 client. Comment
out this test-method, and it can be enabled by dev-engineers
who wish to verify perf changes thru CI jobs.

This method is provided so that one can uncomment it out and
delete the calls to other tests in this file, and get a faster
turnaround on performance validation thru CI-jobs (in the future).

Initial performance benchmarking on Linux-VM (running on a Mac)
for the client-server RPC-msg exerciser shows that:

- For single-client, overhead of L3-logging is ~10-13%
  (This drop needs further analysis.)

- For multiple clients, overhead of L3-logging is <3%,
  which is what is expected.