[Perf] Revise 'time-consumed' tracking on client- & server-side.
This commit simplifies the time-tracking fields for each
client. Previously, we tried to track two different times:

 1. CPU-clock-time for implementing the actual op
 2. Real-time-clock-time for the whole process including
    the overheads of logging.

The idea was to measure (1) [ w/o logging ] and to track it
as one of the arguments to the subsequent l3_log() call.
Another attempt made was to convert the cumulative [2] time
to a throughput metric (num-ops / sec). But trying to mix these
two time-tracking schemes, and jostle between clock_gettime()
calls got confusing and needlessly unreliable.

This commit simplifies everything as follows.

Implement the code as:

  a) clock_gettime() ◁----------------------------+
  b) Do the n-operations in a forever-loop        |--> time measured for
  c) Log, if needed [ Absent in baseline runs ]   | configured clock
  d) clock_gettime() ◁----------------------------+

This commit reworks the time-tracking fields for each side
of the transmission.  We just use the elapsed time between (d) and (a)
to compute the overall throughput.

Client: Get CLOCK_REALTIME for the entire n-messages
  and generate client-side throughput. This will measure
  the impact of L3-logging on real-time throughput.

  When client-exits, report back its throughput to the server.

Server: Use CLOCK_REALTIME as default clock.
  Start timer before forever-loop waiting for messsges from
  the clients. End the timer when all clients exit.

  Assuming that clients are continually pumping messages, this
  is a reasonable way to measure overall elapsed-time.

  Server will report avg throughput across all clients.

The server program has already been extended to support cmdline
flags to select the type of clock to use.

So, with this support, one can run the server as
`svmsg_file_server --clock-process-cputime-id` to run through the
u-benchmarking using the CLOCK_PROCESS_CPUTIME_ID clock, to get
a read-into the timing metrics using that clock. Default clock
is CLOCK_REALTIME.

test.sh: Driver script is enhanced to run various perf-bm tests, varying
  the run-parameters. See `./test.sh --help` for detailed usage info.
  Fix shellcheck errors.

- Added svr_clock_calibrate() & svr_clock_overhead() to run thru
  clock calibration on test machine.

- Improved / expanded msgs reporting metrics / summary.

With this rework, all perf-test-methods will report the client-
and server-side throughput using real-time elapsed, aggregating
across c-clients and n-messages.

Tests can be run using, say, --clock-process-cputime-id flag,
to generate the overhead due to logging in terms of CPU-time
spent when L3 is enabled. Throughput reported on server-side is
also scaled based on CPU-time-ns elapsed.

Fix for debug builds.