[Perf] Relocate svr_clock_calibrate() to end of server-code.
Empirical perf-results on Linux-workstation show that the
drop with L3-Logging enabled is more than the expected ~1-2%.
Commenting out the call to svr_clock_calibrate() has shown
that this gap reduces considerably. Theory is that as the
calibration function runs a pair of 1 Mil calls to clock_gettime(),
it may have 'induced' the CPU-clocking to go into some execution
modes that resulted in performance skew, downstream.

- To remove any doubts that clock-instrumentation is perturbing
  overall perf & metrics, relocate svr_clock_calibrate() to the
  end of server main(), after all workload and perf-metrics
  collection is done-done.

This way, we still get the calibration numbers when needed on any
user's test machine, and still avoid potential perturbations to
perf results.