[Perf] Refactor svmsg_file_server.c to prepare for threaded-server execution.
This commit lays down the ground work to run the server program of the
client-server performance u-benchmarking toolkit as multi-threaded execution.
The goal is to invoke and measure the overhead of concurrent logging
performed by multiple server threads in response to messages received from
n-concurrent clients.

- Add support for --num-server-threads <n> argument, but currently not invoked.
  Default is num_threds = 1

- Break-out inline code into svr_op_*() work-horse functions to implement
  the important msg-types. Driver fn is svr_proc_process_msg(), which is
  currently invoked by a single server-thread.

These changes lay-down enough pthread-create/join code to make it easier
to run, stabilize and re-benchmark overheads of logging using a multi-threaded
server.