Collapse l3_log() and l3_log_fast() using weak symbols.
We make the C implementation of l3_log() use a weak symbol, and rename
the assembly version of the routine to l3_log(). This way, if the assembly
version is present, we'll pick that up, and the caller doesn't need to
worry about which version to use.

(I'm still not convinced the assembly version is sufficiently faster than the
C version to make all this worthwhile, but ...)